Reviewed-on: https://gitea.kood.tech/georgiisenotrusov/movies-api/pulls/15

+++ internal/service/genre.go
@@ -4,38 +4,47 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"movies-api/internal/errs"
 	"movies-api/internal/models"
 	"movies-api/internal/repository"
+
+	"github.com/go-playground/validator/v10"
 )
 
 // service layer for genres
 type GenreService struct {
-	repo *repository.Repo
+	repo     *repository.Repo
+	validate *validator.Validate
 }
 
 // package prepared for submission
 type GenreSubmission struct {
-	Name string `json: name`
+	Name string `json:"name" validate:"required"`
 }
 
 type GenrePatch struct {
-	Name *string `json: name`
+	Name *string `json:"name"`
 }
 
 // initialize genre service
-func NewGenreService(r *repository.Repo) *GenreService {
-	return &GenreService{repo: r}
+func NewGenreService(r *repository.Repo, v *validator.Validate) *GenreService {
+	return &GenreService{repo: r, validate: v}
 }
 
 // validate the data and call the repo lvl
 func (gs *GenreService) AddGenre(ctx context.Context, sub GenreSubmission) (models.Genre, error) {
+
+	if err := gs.validate.Struct(sub); err != nil {
+		return models.Genre{}, fmt.Errorf("%w: %w", errs.ErrInvalidUserInput, err)
+	}
+
 	//pack data in
 	NewGenre := models.Genre{
 		Name: sub.Name,
 	}
 
 	//call repo lvl
-	g, err := gs.repo.CreateGenre(ctx, NewGenre)
+	g, err := gs.repo.AddGenre(ctx, NewGenre)
 
 	return g, err
 }
@@ -51,22 +60,24 @@ func (gs *GenreService) GetGenre(ctx context.Context, id int64) (models.Genre, e
 	return gs.repo.GetGenre(ctx, id)
 }
 
-func (gs *GenreService) PatchGenre(ctx context.Context, id int64, patch GenrePatch) (models.Genre, error) {
-	g := models.Genre{ID: id}
+func (gs *GenreService) PatchGenre(ctx context.Context, id int64, patch GenrePatch) (models.GenreSummary, error) {
+	g := models.GenreSummary{ID: id}
 	if patch.Name != nil {
 		g.Name = *patch.Name
 	}
 
 	g, err := gs.repo.PatchGenre(ctx, g)
 	if err != nil {
-		return models.Genre{}, err
+		return models.GenreSummary{}, err
 	}
 
 	return g, nil
 }
 
-func (gs *MovieService) DeleteGenre(ctx context.Context, id int64) error {
+func (gs *MovieService) DeleteGenre(ctx context.Context, id int64, force bool) error {
+	if !force {
+		return errs.ErrForce
+	}
 	err := gs.repo.DeleteGenre(ctx, id)
-	fmt.Println(err)
 	return err
 }
